Extracting DbTester to be used in new FE test framework
